Add BytesEq and BytesOrd classes
It became apparent that a custom Eq/Ord class would be useful, as we
need that functionality for normalization, but the naive functions for
floats were too problematic.

Thus, BytesEq and BytesOrd were born. For each boolean function,
(e.g. (.=.)) we include asymmetric version (resp. (.=) and (=.)) to
make it more convenient to compare numeric literals. We also change
the previous BytesNum operators (e.g. (|+|)) to match the new operators
(resp. (.+.)). The reason is that extending the previous syntax to
inequalities (e.g. (|>=)) looked hideous with ligatures, so the
vertical bars were replaced by dots.