Significant Refactoring + Bug Fixes
1. The most major change is in how the ByteSize type is tracked so that it
can be recovered at runtime. Previously it was part of the Bytes type,
a GADT which had a constructor for each size. We have now switched to
a singleton approach, making a singleton for the ByteSize along with a
typeclass for recovering this type at runtime. This cleans up the code
in that we no longer have multiple constructors for Bytes. Even better,
we can now implement Applicative and Monad for general (Bytes s n),
which was previously impossible.

We have also made singleton types for ByteDirection.

2. Multiple normalization bugs were fixed:
   - Negative values were not handled; entered an infinite loop. These
     are now handled correctly.
   - '> 999' comparisons switched '>= 1000', the latter being the
     correct semantics.

3. BytesNum class removed. The many operators was confusing, and
implementing 'Num' lets us use numeric literals, which is nice. The are
are two main drawbacks to 'Num'.
   - Multiplication is non-sensical (units should be squared), but at
     least scalar multiplication is reasonable, so we can close our
     eyes and pretend.
   - For types like AnyBytes that hide the size, we have to arbitrarily
     choose a type when using fromInteger. This means we have to be
     careful when doing comparison, as (MkAnySize 1000) < 2000 will
     return false if the former is > Bytes (fromInteger gives 2000 will
     the type 'MkAnySize (MkBytes B) Integer'). This is why the
     ScalarOrd class was added, to sidestep these tricky Num/Eq+Ord
     interactions.

4. New Div class added so that we can abstract over division in
Conversion / Normalize without forcing users to choose e.g. Integral or
Fractional.

5. With singletons, we have no need for the ad hoc LiftBase classes.
Instead, we have a new Isomorphism class that, due to its more
principled nature, allows us to implement lifting functions
generically.

6. Added more tests. We also added the runtime arg "--max-runs" that
allows to pass in a global flag if we want to modify the number of
successful hedgehog runs per test.