Replace Num with custom Math hierarchy
It would be really nice if we could split Bytes and NetworkBytes up,
e.g., by putting their typeclass instances into the typeclasses
themselves. This would improve readability as all the logically similar
code would be in the same place.

There is a problem. The AnyX types use Normalize in their Num instances,
and Normalize relies on the rest of the typeclass zoo. Thus we can't
split out this functionality, unless we:

1. Live with an orphan instance for Num.
2. Remove normalize from Num instances.
3. Replace Num with a class we define ourselves.

One seems ill-motivated at this stage. Two is something we may want to
do in the future for performance reasons, but there are several
advantages to going with three.

1. Multiplication for our byte types is technically nonsense (We can
   cover our eyes, but it's still a tad gross).
2. Num's fromInteger played a role in a normalization bug where, e.g.,

       (MkAnySize SGB (MkBytes 50)) < 1_000

   returned false, because the literal on RHS was being transformed to
   (MkAnySize SB (MkBytes 1_000)). This is clearly the result of
   insufficient attention, but it is still an easy mistake to make.
   Being able to make comparisons with literals is nice, but we can do
   that with the ScalarOrd class without worrying about fromInteger
   interactions.
3. The Num hierarchy is a bit of a mess. Making our own typeclasses
   based on abstract algebra gives us a clean, organized interface.

Thus we have gone with three, implementing our own algebraic structures,
keeping in mind that we will want to investigate two at a later date.