Make 'next (PB)' and 'prev(B)' type errors
There is a tradeoff between how to handle attempting to increase the max
byte size or decrease the min byte size. Up until now, they have essentially
been id. This helps ease-of-use (don't have to prove your type works with
next because it works for all sizes) and makes the code nicer because
we just have to write a single instance.

On the other hand, this behavior can hide errors, and in fact this happened
on one occasion. Normalization on negative numbers was entering an infinite
loop (continually calling next) since it failed a basic 'b < 1 -> next b'
check. Obviously this was a logic error, but it would have been prevented
had we disallowed increasing/decreasing the byte size indefinitely. We have
now implemented these errors, omitting IncByteSize and DecByteSize instances
for the max, min units, respectively.

This also has the nice benefit that we can make our type families injective,
increasing safety and potentially improving type inference.