Improve docs, remove some ScalarX instances
The ScalarX instances have been removed from classes that existentially
quantify the size. This change was motivated by the fact that such types
defined equality in terms of equivalence classes, and the ScalarX
functions break the substitution law.

This is especially problematic when combined with other functions that
rely on this equivalence relation to be lawful, e.g. normalize.

While we still provide the ability to break this law (e.g. show,
exposing the underlying value/type) providing the functions that
actually do so makes it too easy for users to accidentally do the
wrong thing. In this way, we trade some functionality (of dubious
value), for a safer API.