Rename convert fns for RequiredTypeArguments
With RequiredTypeArguments, we arrived at the API we always wanted i.e.
explicitly passing in the required type. This is only available with
GHC 9.10, however, so we guard it with CPP.

Because we want this new function to get priority, we give it the better
name, 'covert'. We thus rename the original 'convert' to 'convert_'.

We also choose slightly better names for the low-level convert functions.